+++ app.py
@@ -22,82 +22,24 @@ def thankyou():
 	return render_template("thankyou.html")
 
 
-# ========================資料庫連線測試===================================
-# AWS
-# Use this code snippet in your app.
-# If you need more information about configurations or implementing the sample code, visit the AWS docs:   
-# https://aws.amazon.com/developers/getting-started/python/
-
-import boto3
-import base64
-from botocore.exceptions import ClientError
-
-
-def get_secret():
-
-    secret_name = "conSQL"
-    region_name = "us-east-1"
-
-    # Create a Secrets Manager client
-    session = boto3.session.Session()
-    client = session.client(
-        service_name='secretsmanager',
-        region_name=region_name
-    )
-
-    # In this sample we only handle the specific exceptions for the 'GetSecretValue' API.
-    # See https://docs.aws.amazon.com/secretsmanager/latest/apireference/API_GetSecretValue.html
-    # We rethrow the exception by default.
-
-    try:
-        get_secret_value_response = client.get_secret_value(
-            SecretId=secret_name
-        )
-    except ClientError as e:
-        if e.response['Error']['Code'] == 'DecryptionFailureException':
-            # Secrets Manager can't decrypt the protected secret text using the provided KMS key.
-            # Deal with the exception here, and/or rethrow at your discretion.
-            raise e
-        elif e.response['Error']['Code'] == 'InternalServiceErrorException':
-            # An error occurred on the server side.
-            # Deal with the exception here, and/or rethrow at your discretion.
-            raise e
-        elif e.response['Error']['Code'] == 'InvalidParameterException':
-            # You provided an invalid value for a parameter.
-            # Deal with the exception here, and/or rethrow at your discretion.
-            raise e
-        elif e.response['Error']['Code'] == 'InvalidRequestException':
-            # You provided a parameter value that is not valid for the current state of the resource.
-            # Deal with the exception here, and/or rethrow at your discretion.
-            raise e
-        elif e.response['Error']['Code'] == 'ResourceNotFoundException':
-            # We can't find the resource that you asked for.
-            # Deal with the exception here, and/or rethrow at your discretion.
-            raise e
-    else:
-        # Decrypts secret using the associated KMS key.
-        # Depending on whether the secret is a string or binary, one of these fields will be populated.
-        if 'SecretString' in get_secret_value_response:
-            secret = get_secret_value_response['SecretString']
-        else:
-            decoded_binary_secret = base64.b64decode(get_secret_value_response['SecretBinary'])
-            
-    # Your code goes here. 
-
-result=get_secret()
-print("結果:",result)
-
+#=============================================================
 
 import mysql.connector
+import boto3
+import json
+
+client=boto3.client('secretsmanager')
+response=client.get_secret_value(SecretId='conSQL')
+
+secretDict=json.loads(response['SecretString'])
 mydb = mysql.connector.connect(
-    host="database-1.cohxynft1tdv.us-east-1.rds.amazonaws.com",
-    user=result[username],
-    password=result[password],
-    database="travelSite"
+    host=secretDict['host'],
+    user=secretDict['username'],
+    password=secretDict['password'],
+    database=secretDict['dbname']
 )
 print(mydb)
 mycursor=mydb.cursor()
-#=========================================測試結束=====================================
 
 #========================================原版=========================================
