+++ aws-testSQL.js
@@ -19,3 +19,10 @@ const main =async(event) =>{
 }
 
 exports.handler = main
+
+
+
+
+
+
+
